app/requirements.txt
fastapi
uvicorn[standard]
google-cloud-secret-manager
google-auth
google-generativeai
requests